raise/src/main/webapp/app/entities/demand/demand.component.ts
import { mixins } from 'vue-class-component';

import { Component, Inject } from 'vue-property-decorator';
import Vue2Filters from 'vue2-filters';
import { IDemand } from '@/shared/model/demand.model';
import AlertMixin from '@/shared/alert/alert.mixin';

import DemandService from './demand.service';

@Component
export default class Demand extends mixins(Vue2Filters.mixin, AlertMixin) {
  @Inject('demandService') private demandService: () => DemandService;
  private removeId: number = null;
  public itemsPerPage = 20;
  public queryCount: number = null;
  public page = 1;
  public previousPage = 1;
  public propOrder = 'id';
  public reverse = false;
  public totalItems = 0;

  public demands: IDemand[] = [];

  public isFetching = false;

  public mounted(): void {
    this.retrieveAllDemands();
  }

  public clear(): void {
    this.page = 1;
    this.retrieveAllDemands();
  }

  public retrieveAllDemands(): void {
    this.isFetching = true;

    const paginationQuery = {
      page: this.page - 1,
      size: this.itemsPerPage,
      sort: this.sort()
    };
    this.demandService()
      .retrieve(paginationQuery)
      .then(
        res => {
          this.demands = res.data;
          this.totalItems = Number(res.headers['x-total-count']);
          this.queryCount = this.totalItems;
          this.isFetching = false;
        },
        err => {
          this.isFetching = false;
        }
      );
  }

  public prepareRemove(instance: IDemand): void {
    this.removeId = instance.id;
    if (<any>this.$refs.removeEntity) {
      (<any>this.$refs.removeEntity).show();
    }
  }

  public removeDemand(): void {
    this.demandService()
      .delete(this.removeId)
      .then(() => {
        const message = this.$t('raiseApp.demand.deleted', { param: this.removeId });
        this.alertService().showAlert(message, 'danger');
        this.getAlertFromStore();
        this.removeId = null;
        this.retrieveAllDemands();
        this.closeDialog();
      });
  }

  public sort(): Array<any> {
    const result = [this.propOrder + ',' + (this.reverse ? 'asc' : 'desc')];
    if (this.propOrder !== 'id') {
      result.push('id');
    }
    return result;
  }

  public loadPage(page: number): void {
    if (page !== this.previousPage) {
      this.previousPage = page;
      this.transition();
    }
  }

  public transition(): void {
    this.retrieveAllDemands();
  }

  public changeOrder(propOrder): void {
    this.propOrder = propOrder;
    this.reverse = !this.reverse;
    this.transition();
  }

  public closeDialog(): void {
    (<any>this.$refs.removeEntity).hide();
  }
}
